refactor code to mvc model resolve #2
