+++ src/index.js
@@ -2,10 +2,7 @@ const express = require('express')
 const cors = require('cors')
 const bodyParser = require('body-parser')
 require('./db/mongoose')
-const usersRouter = require('./routers/users')
-const tasksRouter = require('./routers/tasks')
-const projectsRouter = require('./routers/projects')
-const testRouter = require('./route/routers')
+const Router = require('./router/routers')
 const app = express()
 const HOST = '127.0.0.1'
 const PORT = process.env.PORT || 44010
@@ -13,5 +10,5 @@ const PORT = process.env.PORT || 44010
 app.use(express.json())
 app.use(bodyParser.json())
 app.use(cors())
-app.use(usersRouter, tasksRouter, projectsRouter, testRouter)
+app.use(Router)
 app.listen(PORT,HOST, () => console.log(`Example app listening on http://${HOST}:${PORT}`))
